V2.0
This version is capable to crawled all the news pages on JLU website. Besides, it is able to export the relative files of the data of news-released time by department(data.csv) and word frequency (based on jieba technology, frequency.csv). 
It is also worth mention that I am fucking tired, I should not do all these works on my own.